BREAKING(crypto): clean up module
